backend/requirements.txt
# --- Core Django & ASGI ---
Django==5.0.3
asgiref==3.11.1
daphne==4.1.0
channels==4.0.0
channels-redis==4.2.0
psycopg2-binary==2.9.9
redis==5.0.3
django-redis==6.0.0
celery==5.3.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.2
PyJWT==2.12.1
google-auth==2.29.0
google-auth-oauthlib==1.2.0
django-jazzmin==3.0.0
django-tinymce==3.6.1
bleach==6.3.0
django-storages==1.14.2
boto3==1.34.69
Pillow==10.2.0
protobuf>=4.25.3,<6.0.0dev
python-dotenv==1.0.1
pywebpush==1.14.0
Requests==2.31.0
sentry-sdk==1.43.0
tweepy==4.14.0
newspaper3k==0.2.8
whitenoise==6.6.0
lxml_html_clean
google-generativeai==0.8.6